# docs/v3/openapi3.json
{
	"components": {
		"schemas": {
...
			"datastore.PaginationData": {
				"properties": {
					"has_next_page": {
						"type": "boolean"
					},
					"has_prev_page": {
						"type": "boolean"
					},
					"next_page_cursor": {
						"type": "string"
					},
					"per_page": {
						"type": "integer"
					},
					"prev_page_cursor": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"handlers.Stub": {
				"type": "object"
			},
			"httpheader.HTTPHeader": {
				"additionalProperties": {
					"items": {
						"type": "string"
					},
					"type": "array"
				},
				"type": "object"
			},
...
			"models.EndpointResponse": {
				"properties": {
					"advanced_signatures": {
						"type": "boolean"
					},
					"authentication": {
						"$ref": "#/components/schemas/datastore.EndpointAuthentication"
					},
					"created_at": {
						"type": "string"
					},
					"deleted_at": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"events": {
						"type": "integer"
					},
					"http_timeout": {
						"type": "integer"
					},
					"name": {
						"type": "string"
					},
					"owner_id": {
						"type": "string"
					},
					"project_id": {
						"type": "string"
					},
					"rate_limit": {
						"type": "integer"
					},
					"rate_limit_duration": {
						"type": "integer"
					},
					"secrets": {
						"items": {
							"$ref": "#/components/schemas/datastore.Secret"
						},
						"type": "array"
					},
					"slack_webhook_url": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/datastore.EndpointStatus"
					},
					"support_email": {
						"type": "string"
					},
					"uid": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"url": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"models.EventResponse": {
				"properties": {
					"acknowledged_at": {
						"type": "string"
					},
					"app_id": {
						"description": "Deprecated",
						"type": "string"
					},
					"created_at": {
						"type": "string"
					},
					"data": {
						"description": "Data is an arbitrary JSON value that gets sent as the body of the\nwebhook to the endpoints",
						"items": {
							"type": "integer"
						},
						"type": "array"
					},
					"deleted_at": {
						"type": "string"
					},
					"endpoint_metadata": {
						"items": {
							"$ref": "#/components/schemas/datastore.Endpoint"
						},
						"type": "array"
					},
					"endpoints": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"event_type": {
						"type": "string"
					},
					"headers": {
						"$ref": "#/components/schemas/httpheader.HTTPHeader"
					},
					"idempotency_key": {
						"type": "string"
					},
					"is_duplicate_event": {
						"type": "boolean"
					},
					"project_id": {
						"type": "string"
					},
					"raw": {
						"type": "string"
					},
					"source_id": {
						"type": "string"
					},
					"source_metadata": {
						"$ref": "#/components/schemas/datastore.Source"
					},
					"uid": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					},
					"url_query_params": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"models.PagedResponse": {
				"properties": {
					"content": {},
					"pagination": {
						"$ref": "#/components/schemas/datastore.PaginationData"
					}
				},
				"type": "object"
			},
...
			"util.ServerResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "boolean"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {
			"ApiKeyAuth": {
				"in": "header",
				"name": "Authorization",
				"type": "apiKey"
			}
		}
	},
	"info": {
		"contact": {
			"email": "support@getconvoy.io",
			"name": "Convoy Support",
			"url": "https://getconvoy.io/docs"
		},
		"description": "Convoy is a fast and secure webhooks proxy. This document contains datastore.s API specification.",
		"license": {
			"name": "Mozilla Public License 2.0",
			"url": "https://www.mozilla.org/en-US/MPL/2.0/"
		},
		"termsOfService": "https://getconvoy.io/terms",
		"title": "Convoy API Reference",
		"version": "24.1.4"
	},
	"openapi": "3.0.3",
	"paths": {
...
		"/v1/projects/{projectID}/endpoints": {
			"get": {
				"description": "This endpoint fetches an endpoints",
				"operationId": "GetEndpoints",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"schema": {
							"enum": [
								"next",
								"prev"
							],
							"type": "string"
						},
						"x-enum-varnames": [
							"Next",
							"Prev"
						]
					},
					{
						"description": "A pagination cursor to fetch the next page of a list",
						"in": "query",
						"name": "next_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The owner ID of the endpoint",
						"in": "query",
						"name": "ownerId",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The number of items to return per page",
						"in": "query",
						"name": "perPage",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "A pagination cursor to fetch the previous page of a list",
						"in": "query",
						"name": "prev_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The name of the endpoint",
						"in": "query",
						"name": "q",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order, values are `ASC` or `DESC`, defaults to `DESC`",
						"in": "query",
						"name": "sort",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"allOf": [
														{
															"$ref": "#/components/schemas/models.PagedResponse"
														},
														{
															"properties": {
																"content": {
																	"items": {
																		"$ref": "#/components/schemas/models.EndpointResponse"
																	},
																	"type": "array"
																}
															},
															"type": "object"
														}
													]
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "List all endpoints",
				"tags": [
					"Endpoints"
				]
			},
			"post": {
				"description": "This endpoint creates an endpoint",
				"operationId": "CreateEndpoint",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.CreateEndpoint"
							}
						}
					},
					"description": "Endpoint Details",
					"required": true,
					"x-originalParamName": "endpoint"
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/models.EndpointResponse"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Create an endpoint",
				"tags": [
					"Endpoints"
				]
			}
		},
...
		"/v1/projects/{projectID}/events": {
			"get": {
				"description": "This endpoint fetches app events with pagination",
				"operationId": "GetEventsPaged",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "direction",
						"schema": {
							"enum": [
								"next",
								"prev"
							],
							"type": "string"
						},
						"x-enum-varnames": [
							"Next",
							"Prev"
						]
					},
					{
						"description": "The end date",
						"in": "query",
						"name": "endDate",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "A list of endpoint ids to filter by",
						"in": "query",
						"name": "endpointId",
						"schema": {
							"items": {
								"type": "string"
							},
							"type": "array"
						}
					},
					{
						"description": "IdempotencyKey to filter by",
						"in": "query",
						"name": "idempotencyKey",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "A pagination cursor to fetch the next page of a list",
						"in": "query",
						"name": "next_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The number of items to return per page",
						"in": "query",
						"name": "perPage",
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "A pagination cursor to fetch the previous page of a list",
						"in": "query",
						"name": "prev_page_cursor",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Any arbitrary value to filter the events payload",
						"in": "query",
						"name": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort order, values are `ASC` or `DESC`, defaults to `DESC`",
						"in": "query",
						"name": "sort",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "A list of Source IDs to filter the events by.",
						"in": "query",
						"name": "sourceId",
						"schema": {
							"items": {
								"type": "string"
							},
							"type": "array"
						}
					},
					{
						"description": "The start date",
						"in": "query",
						"name": "startDate",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"allOf": [
														{
															"$ref": "#/components/schemas/models.PagedResponse"
														},
														{
															"properties": {
																"content": {
																	"items": {
																		"$ref": "#/components/schemas/models.EventResponse"
																	},
																	"type": "array"
																}
															},
															"type": "object"
														}
													]
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "OK"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "List all events",
				"tags": [
					"Events"
				]
			},
			"post": {
				"description": "This endpoint creates an endpoint event",
				"operationId": "CreateEndpointEvent",
				"parameters": [
					{
						"description": "Project ID",
						"in": "path",
						"name": "projectID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/models.CreateEvent"
							}
						}
					},
					"description": "Event Details",
					"required": true,
					"x-originalParamName": "event"
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Bad Request"
					},
					"401": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Unauthorized"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/util.ServerResponse"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/handlers.Stub"
												}
											},
											"type": "object"
										}
									]
								}
							}
						},
						"description": "Not Found"
					}
				},
				"security": [
					{
						"ApiKeyAuth": []
					}
				],
				"summary": "Create an event",
				"tags": [
					"Events"
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://dashboard.getconvoy.io/api"
		}
	],
...
}